backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
requests==2.32.4
langchain
langchain-openai
langchain-community
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
openai 